// frontend/src/baml_client/types.ts
export interface Analysis {
  keyAssumptions: Record<string, number | string>
  metric_type: "proforma" | "monthly"
  proformaMetrics?: ProformaMetrics[] | null
  monthlyMetrics?: MonthlyComparison[] | null
  formulas: Formula[]
  
}
// ...
export interface Deltas {
  arr: number
  revenue: number
  netIncome: number
  bankBalance: number
  
}

export interface EBITDA {
  depreciation: number
  amortization: number
  interestExpense: number
  ebitdaValue: number
  ebitdaMargin: number
  
}

export interface Expenses {
  engineering: number
  generalAndAdmin: number
  salesAndMarketing: number
  total: number
  
}

export interface Experience {
  role: string
  company: string
  
}

export interface FinancialMetrics {
  arr: number
  revenue: number
  costOfRevenue: CostOfRevenue
  grossProfit?: GrossProfit | null
  expenses: Expenses
  netIncome: number
  cashFlow: CashFlow
  bankBalance: number
  
}

export interface FinancialMetricsEBITDA {
  year: string
  revenue?: Revenue | null
  costs?: Costs | null
  profitability?: Profitability | null
  perShare?: PerShare | null
  ebitda?: EBITDA | null
  
}

export interface Formula {
  name: string
  purpose: string
  parameters: string[]
  expression: string
  
}

export interface GrossProfit {
  amount: number
  margin: number
  
}

export interface MissingParam {
  name: string
  formula: Formula
  
}

export interface MonthlyComparison {
  month: string
  target?: FinancialMetrics | null
  actual: FinancialMetrics
  deltas: Deltas
  
}

export interface Name {
  first: string
  last: string
  
}

export interface PerShare {
  basic: number
  diluted: number
  
}

export interface Profitability {
  grossMargin: number
  operatingIncome: number
  otherIncome: number
  incomeTaxes: number
  netIncome: number
  
}

export interface ProformaMetrics {
  period: string
  effectiveGrossRevenue: number
  operatingExpenses: number
  netOperatingIncome?: number | null
  capitalExpenditures?: number | null
  cashFlowFromOperations?: number | null
  capRate?: number | null
  propertyValue?: number | null
  valuePerUnit?: number | null
  yieldOnCost?: number | null
  totalProjectCost?: number | null
  developmentSpread?: number | null
  
}